ammount of sessions read from server

+++ Hackovid/app/src/main/java/com/apptest/nensalparc/Hour.kt
@@ -1,6 +1,6 @@
 package com.apptest.nensalparc
 
 data class HourModel (
-    val start: Long? = null,
+    val start: Int? = null,
     val timeFractions: ArrayList<TimeFractionModel>? = null
 )
